refactor(web): extract connection form validation to separate module
Moves validation logic from ConnectionForm.tsx into a dedicated
connectionFormValidation.ts module to prepare for adding Bridge, VLAN,
and other connection types.

The previous approach had inline type checks (formValues.type ===
ConnectionType.BOND) scattered throughout validateConnectionForm, which
would not scale well. The new structure uses type-specific validator
functions (validateBondFields, validateCommonFields) that are composed
together, making it straightforward to add new types.

Also improves bond options validation to use sprintf and formatList for
proper i18n of the mode names list.
